requirements.txt
hydra-core
wandb
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
scipy
pytorch-lightning
timm
transformers
omegaconf
pillow
tensorboard
loguru
pre-commit
black
flake8
isort
torchmetrics